+++ CHANGELOG.md
@@ -24,6 +24,9 @@
 * Rewrite `die` implementation to use `die` from `base`.
 * [#19](https://github.com/kowainik/relude/issues/19):
   Rewrite `.hlint.yaml` to Dhall.
+* Move `stdin`- and `stdout`-related functions to new module `Relude.Lifted.Terminal`.
+* [#67](https://github.com/kowainik/relude/issues/67):
+  Add HLint rules for `put*` functions.
 
 0.2.0
 =====